Move colima_host_ip() to runtime/colima.py
This is a better home for a low-level Colima networking helper than
setup.py (an orchestration module). Updates imports in github_token.py
and provisioning.py.
